+++ pipeline/config/targets/build-global-insights.ts
@@ -39,4 +39,5 @@ export default [
   'mtfry', // meimon-taiyo-ferry
   'itkfry', // itsukishima-kisen
   'kytbus', // kyoto-bus
+  'od9bus', // odakyu-bus
 ];