+++ pipeline/config/targets/build-insights.ts
@@ -42,4 +42,5 @@ export default [
   'mtfry', // meimon-taiyo-ferry
   'itkfry', // itsukishima-kisen
   'kytbus', // kyoto-bus
+  'od9bus', // odakyu-bus
 ];